Move test utils out of TweedledumBase (on their way to their own file eventually).
